Reorganize core.task package
task.tracker did not contain tracking code so I split it into "jobs"
and "update".
