+++ src/main/scala/mesosphere/marathon/core/CoreModule.scala
@@ -6,7 +6,7 @@ import mesosphere.marathon.core.launchqueue.LaunchQueueModule
 import mesosphere.marathon.core.leadership.LeadershipModule
 import mesosphere.marathon.core.plugin.PluginModule
 import mesosphere.marathon.core.task.bus.TaskBusModule
-import mesosphere.marathon.core.task.tracker.TaskTrackerModule
+import mesosphere.marathon.core.task.jobs.TaskJobsModule
 
 /**
   * The exported interface of the [[CoreModuleImpl]].
@@ -17,7 +17,7 @@ import mesosphere.marathon.core.task.tracker.TaskTrackerModule
 trait CoreModule {
   def leadershipModule: LeadershipModule
   def taskBusModule: TaskBusModule
-  def taskTrackerModule: TaskTrackerModule
+  def taskJobsModule: TaskJobsModule
   def launcherModule: LauncherModule
   def appOfferMatcherModule: LaunchQueueModule
   def pluginModule: PluginModule
