+++ src/main/scala/mesosphere/marathon/core/task/update/TaskStatusUpdateProcessor.scala
@@ -1,4 +1,4 @@
-package mesosphere.marathon.core.task.tracker
+package mesosphere.marathon.core.task.update
 
 import org.apache.mesos.Protos.TaskStatus
