+++ src/test/scala/mesosphere/marathon/core/task/update/impl/steps/NotifyLaunchQueueStepImplTest.scala
@@ -1,4 +1,4 @@
-package mesosphere.marathon.core.task.tracker.impl.steps
+package mesosphere.marathon.core.task.update.impl.steps
 
 import mesosphere.marathon.MarathonTestHelper
 import mesosphere.marathon.core.launchqueue.LaunchQueue
