+++ src/test/scala/mesosphere/marathon/core/task/update/impl/steps/UpdateTaskTrackerStepImplTest.scala
@@ -1,4 +1,4 @@
-package mesosphere.marathon.core.task.tracker.impl.steps
+package mesosphere.marathon.core.task.update.impl.steps
 
 import mesosphere.marathon.MarathonSchedulerDriverHolder
 import mesosphere.marathon.Protos.MarathonTask
